Introduce the Container Metadata Store
This is a central storage for container-related metadata, and is
intended to replace the standalone collectors in Autodiscovery and the
tagger. This commit is a WIP, and contains the main interfaces and a
proof of concept implementation of the work that needs to be done in the
tagger to replace the standalone collectors with this store.